+++ src/middlewared/middlewared/plugins/crypto_/certificate_authorities.py
@@ -1,7 +1,6 @@
 import random
 
-
-from middlewared.schema import accepts, Bool, Dict, Int, Patch, Ref, returns, Str
+from middlewared.schema import accepts, Bool, Dict, Int, Patch, Str
 from middlewared.service import CRUDService, private, ValidationErrors
 import middlewared.sqlalchemy as sa
 
@@ -9,8 +8,7 @@ from .cert_entry import get_ca_result_entry
 from .common_validation import _validate_common_attributes, validate_cert_name
 from .dependencies import check_dependencies
 from .utils import (
-    get_cert_info_from_data, _set_required,
-    CA_TYPE_EXISTING, CA_TYPE_INTERNAL, CA_TYPE_INTERMEDIATE, CERT_TYPE_INTERNAL
+    get_cert_info_from_data, _set_required, CA_TYPE_EXISTING, CA_TYPE_INTERNAL, CA_TYPE_INTERMEDIATE
 )
 
 
@@ -265,132 +263,6 @@ class CertificateAuthorityService(CRUDService):
         return await self.get_instance(pk)
 
     @accepts(
-        Dict(
-            'ca_sign_csr',
-            Int('ca_id', required=True),
-            Int('csr_cert_id', required=True),
-            Str('name', required=True),
-            Ref('cert_extensions'),
-            register=True
-        )
-    )
-    @returns(Ref('certificate_entry'))
-    async def ca_sign_csr(self, data):
-        """
-        Sign CSR by Certificate Authority of `ca_id`
-
-        Sign CSR's and generate a certificate from it. `ca_id` provides which CA is to be used for signing
-        a CSR of `csr_cert_id` which exists in the system
-
-        `cert_extensions` can be specified if specific extensions are to be set in the newly signed certificate.
-
-        .. examples(websocket)::
-
-          Sign CSR of `csr_cert_id` by Certificate Authority of `ca_id`
-
-            :::javascript
-            {
-                "id": "6841f242-840a-11e6-a437-00e04d680384",
-                "msg": "method",
-                "method": "certificateauthority.ca_sign_csr",
-                "params": [{
-                    "ca_id": 1,
-                    "csr_cert_id": 1,
-                    "name": "signed_cert"
-                }]
-            }
-        """
-        return await self.__ca_sign_csr(data)
-
-    @accepts(
-        Ref('ca_sign_csr'),
-        Str('schema_name', default='certificate_authority_update')
-    )
-    async def __ca_sign_csr(self, data, schema_name):
-        verrors = ValidationErrors()
-
-        ca_data = await self.query([('id', '=', data['ca_id'])])
-        csr_cert_data = await self.middleware.call('certificate.query', [('id', '=', data['csr_cert_id'])])
-
-        if not ca_data:
-            verrors.add(
-                f'{schema_name}.ca_id',
-                f'No Certificate Authority found for id {data["ca_id"]}'
-            )
-        else:
-            ca_data = ca_data[0]
-            if not ca_data.get('privatekey'):
-                verrors.add(
-                    f'{schema_name}.ca_id',
-                    'Please use a CA which has a private key assigned'
-                )
-
-        if not csr_cert_data:
-            verrors.add(
-                f'{schema_name}.csr_cert_id',
-                f'No Certificate found for id {data["csr_cert_id"]}'
-            )
-        else:
-            csr_cert_data = csr_cert_data[0]
-            if not csr_cert_data.get('CSR'):
-                verrors.add(
-                    f'{schema_name}.csr_cert_id',
-                    'No CSR has been filed by this certificate'
-                )
-            else:
-                if not await self.middleware.call('cryptokey.load_certificate_request', csr_cert_data['CSR']):
-                    verrors.add(
-                        f'{schema_name}.csr_cert_id',
-                        'CSR not valid'
-                    )
-                if not csr_cert_data['privatekey']:
-                    verrors.add(
-                        f'{schema_name}.csr_cert_id',
-                        'Private key not found for specified CSR.'
-                    )
-
-        if verrors:
-            raise verrors
-
-        serial = await self.get_serial_for_certificate(ca_data['id'])
-
-        new_cert = await self.middleware.call(
-            'cryptokey.sign_csr_with_ca',
-            {
-                'ca_certificate': ca_data['certificate'],
-                'ca_privatekey': ca_data['privatekey'],
-                'csr': csr_cert_data['CSR'],
-                'csr_privatekey': csr_cert_data['privatekey'],
-                'serial': serial,
-                'digest_algorithm': ca_data['digest_algorithm'],
-                'cert_extensions': data['cert_extensions']
-            }
-        )
-
-        new_csr = {
-            'type': CERT_TYPE_INTERNAL,
-            'name': data['name'],
-            'certificate': new_cert,
-            'privatekey': csr_cert_data['privatekey'],
-            'signedby': ca_data['id']
-        }
-
-        new_csr_id = await self.middleware.call(
-            'datastore.insert',
-            'system.certificate',
-            new_csr,
-            {'prefix': 'cert_'}
-        )
-
-        await self.middleware.call('service.start', 'ssl')
-
-        return await self.middleware.call(
-            'certificate.query',
-            [['id', '=', new_csr_id]],
-            {'get': True}
-        )
-
-    @accepts(
         Patch(
             'ca_create_internal', 'ca_create_intermediate',
             ('add', {'name': 'signedby', 'type': 'int', 'required': True}),
@@ -512,7 +384,9 @@ class CertificateAuthorityService(CRUDService):
         if data.pop('create_type', '') == 'CA_SIGN_CSR':
             # BEING USED BY OLD LEGACY FOR SIGNING CSR'S. THIS CAN BE REMOVED WHEN LEGACY UI IS REMOVED
             data['ca_id'] = id
-            return await self.__ca_sign_csr(data, 'certificate_authority_update')
+            return await self.middleware.call(
+                'certificateauthority.ca_sign_csr_impl', data, 'certificate_authority_update'
+            )
         else:
             for key in ['ca_id', 'csr_cert_id']:
                 data.pop(key, None)
